pyproject.toml
[project]
name = "qna-medical-referenced"
version = "0.1.0"
description = "Medical Q&A chatbot with RAG pipeline for health screening interpretation"
requires-python = ">=3.12,<3.13"
 dependencies = [
    "fastapi>=0.129.0",
    "openai>=1.0.0",
    "pypdf>=4.0,<6.0",
    "pdfplumber>=0.11.4",
    "pymupdf>=1.24.0",
    "readability-lxml>=0.8.0",
    "pydantic-settings>=2.6.0",
    "uvicorn>=0.40.0",
    "nltk>=3.9.2",
    "httpx>=0.28.1",
    "beautifulsoup4>=4.13.0",
    "markdownify>=0.14.0",
    "trafilatura>=1.12.2",
    "wandb>=0.23.0",
    "chromadb>=0.4.0",
    "bcrypt>=4.0.0",
    "litellm>=1.0.0",
    "requests>=2.32.0",
    "urllib3>=2.0.0,<3.0.0",
    "polars>=1.0.0",
    "sf-hamilton[visualization]>=1.0.0",
    "duckdb>=1.0.0",
    "pyyaml>=6.0.0",
    "deepeval>=3.9.0,<4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
    "pre-commit>=3.0.0",
    "bandit>=1.7.0",
]
 evaluation = [
    "deepeval>=3.9.0,<4.0.0",
    "langchain<1",
    "langchain-community<1",
    "langchain-core<1",
    "langchain-openai<1",
    ]
chunkers = [
    "chonkie[semantic]>=1.0.0",
]
extraction = [
    "camelot-py>=0.12.0",
]
reranking = [
    "sentence-transformers>=3.0.0",
]
test = [
    "qna-medical-referenced[dev,evaluation]",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = ["--strict-markers", "-ra", "--durations=10"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: fast isolated tests (no I/O, no external deps)",
    "integration: tests with DB, filesystem, or service stack",
    "e2e: full end-to-end workflow tests",
    "slow: >1s tests",
    "live_api: requires live Qwen API access",
    "live_openrouter: requires live OpenRouter API access",
    "deepeval: DeepEval integration tests (slow, requires API)",
    "e2e_real_apis: end-to-end tests with real APIs (requires ENABLE_REAL_API_TESTS=1)",
    "network: needs internet access",
    "smoke: critical path tests",
    "serial: cannot run in parallel",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E", "F", "I", "UP", "B", "C4", "DTZ", "T10",
    "ISC", "PIE", "PT", "RUF",
]
ignore = ["E501"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "playwright>=1.58.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.15.1",
]
medical = [
    "spacy>=3.8.0",
]

[tool.uv]
no-build-isolation-package = ["grpcio"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true
strict_optional = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "venv",
    ".venv",
    "build",
    "dist",
]
disable_error_code = [
    "import-untyped",
]

[tool.bandit]
exclude_dirs = ["tests", ".venv", "venv"]
skips = ["B101", "B601"]  # Skip assert_used and paramiko_calls

[[tool.mypy.overrides]]
module = [
    "nltk.*",
    "google.*",
    "deepeval.*",
]
ignore_missing_imports = true
